rename(#20): datasource network, local 패키지 분리
